Load soilgrids

Loads a set of layers (defined by layer_names character-vector) from soilgrids into the soil project (this includes files in the defined path and the R environment, see: new_soil_project).

load_soilgrids(project_path, shape_file, layer_names)

Arguments

project_path
Path to the soil project.

shape_file
Shape file of the study area.

layer_names
Character vector for defining the soilgrids layers to load, e.g "BDRICM_M_250m".